feat(overlays): CH metronome delay adjustable in 0.5s steps
The metronome's Delay stepper now moves in half-second increments
(min 0.5s), and the saved value parses as a float so fractional delays
survive reopening the window. The Settings 'Expected cadence' field
accepts fractional seconds too (IntervalSecs int -> float64 backend-side,
backward-compatible in YAML/JSON).

+++ backend/internal/chchain/matcher_test.go
@@ -20,7 +20,7 @@ func (f *fakeSink) StartExternal(name, category string, dur, _ int, _ time.Time,
 	f.calls = append(f.calls, capture{name, category, dur})
 }
 
-func newMatcher(s Sink, enabled bool, pattern string, interval int) *Matcher {
+func newMatcher(s Sink, enabled bool, pattern string, interval float64) *Matcher {
 	return New(s, func() config.CHChainSettings {
 		return config.CHChainSettings{Enabled: enabled, Pattern: pattern, IntervalSecs: interval}
 	})
